@@ -2,7 +2,7 @@
     "name": "fumikito/never-let-me-go",
     "description": "If someone wants to leave your WordPress, let him go.",
     "type": "plugin",
-    "license": "GPL-3.0",
+    "license": "GPL-3.0-or-later",
     "scripts": {
         "test": "phpunit",
         "cs": [
@@ -19,14 +19,14 @@
     ],
     "minimum-stability": "stable",
     "require": {
-        "php": ">=5.3.0",
+        "php": "^5.6|^7.0",
         "kuno1/redirect-to-woocommerce": "^1.0"
     },
     "require-dev": {
-        "phpunit/phpunit": "4.8.27",
-        "phpunit/php-code-coverage": "2.2.*",
-        "squizlabs/php_codesniffer": "~2.8",
-        "wp-coding-standards/wpcs": "~0.11"
+        "phpunit/phpunit": "^5.7",
+        "squizlabs/php_codesniffer": "~3.3.1",
+        "wp-coding-standards/wpcs": "~2.0",
+        "phpcompatibility/phpcompatibility-wp": "^2.0"
     },
     "autoload": {
         "psr-0": {
