@@ -17,15 +17,15 @@ use NeverLetMeGo\Utility\Input;
  * @property-read string $dir
  * @property-read string $url
  * @property-read string $version
+ * @property-read string[] $meta_to_keep
  */
 class Application extends Singleton {
-
+	
 	/**
 	 * @var string
 	 */
 	protected $name = 'never_let_me_go';
-
-
+	
 	/**
 	 * nonce用に接頭辞をつけて返す
 	 *
@@ -36,7 +36,7 @@ class Application extends Singleton {
 	public function nonce_action( $action ) {
 		return $this->name . '_' . $action;
 	}
-
+	
 	/**
 	 * wp_nonce_fieldのエイリアス
 	 *
@@ -45,7 +45,7 @@ class Application extends Singleton {
 	public function nonce_field( $action ) {
 		wp_nonce_field( $this->nonce_action( $action ), $this->nonce_key );
 	}
-
+	
 	/**
 	 * Delete current_user account
 	 *
@@ -53,13 +53,15 @@ class Application extends Singleton {
 	 */
 	public function delete_current_user() {
 		$user_id = get_current_user_id();
+		
 		return $this->delete_user( $user_id );
 	}
-
+	
 	/**
 	 * Delete user account
 	 *
 	 * @param int $user_id
+	 *
 	 * @return true|\WP_Error
 	 */
 	public function delete_user( $user_id ) {
@@ -68,8 +70,9 @@ class Application extends Singleton {
 		if ( ! $user_id ) {
 			return new \WP_Error( 404, __( 'User doesn\'t exist.', 'never-let-me-go' ) );
 		}
+		$user  = wp_get_current_user();
 		$error = new \WP_Error();
-		if ( current_user_can( 'administrator' ) ) {
+		if ( $user->has_cap( 'activate_plugins' ) ) {
 			$error->add( 400, __( 'You are administrator!', 'never-let-me-go' ) );
 		}
 		/**
@@ -80,10 +83,12 @@ class Application extends Singleton {
 		 * User can't remove their account.
 		 *
 		 * @filter nlmg_validate_user
-		 * @since 1.0.0
+		 *
 		 * @param \WP_Error $error Error object
 		 * @param int $user_id User ID to leave
+		 *
 		 * @return \WP_Error
+		 * @since 1.0.0
 		 */
 		$result = apply_filters( 'nlmg_validate_user', $error, $user_id );
 		// If error is not empty, display page.
@@ -96,33 +101,37 @@ class Application extends Singleton {
 		 *
 		 * Executed before leave site.
 		 *
+		 * @param int $user_id
+		 *
 		 * @since 1.0.0
 		 * @action nlmg_before_leave
-		 * @param int $user_id
 		 */
 		do_action( 'nlmg_before_leave', $user_id );
 		wp_logout();
-		if ( $this->option['keep_account'] ) {
+		if ( $this->option[ 'keep_account' ] ) {
 			delete_user_meta( $user_id, $wpdb->prefix . 'capabilities' );
 			delete_user_meta( $user_id, $wpdb->prefix . 'user_level' );
-			switch ( $this->option['destroy_level'] ) {
+			switch ( $this->option[ 'destroy_level' ] ) {
 				case 0:
 					// Do nothing.
 					break;
 				default:
-					$login = uniqid( 'nlmg.', true );
+					$login = uniqid( 'deleted-', true );
 					$pass  = wp_generate_password( 20 );
 					// Avoid password change mail.
 					add_filter( 'send_password_change_email', '__return_false', 9999 );
 					// Avoid email change mail.
 					add_filter( 'send_email_change_email', '__return_false', 9999 );
 					// Update user table.
-					$user_id = wp_update_user( array(
-						'ID'           => $user_id,
-						'user_pass'    => $pass,
-						'user_email'   => '',
-						'display_name' => __( 'Deleted User', 'never-let-me-go' ),
-					) );
+					$replaced = apply_filters( 'nlmg_hashed_user_data', [
+						'ID'            => $user_id,
+						'user_pass'     => $pass,
+						'user_email'    => $login . '@example.com',
+						'display_name'  => sprintf( __( 'Deleted User #%d', 'never-let-me-go' ), $user_id ),
+						'user_nicename' => $login,
+						'user_url'      => '',
+					] );
+					$user_id = wp_update_user( $replaced );
 					// Update user_login.
 					$wpdb->update(
 						$wpdb->users,
@@ -135,6 +144,10 @@ class Application extends Singleton {
 						array( '%s' ),
 						array( '%d' )
 					);
+					// Mask user meta.
+					if ( $this->filtered_keys() ) {
+						$this->delete_user_meta( $user_id );
+					}
 					/**
 					 * nlmg_after_hashing_user
 					 *
@@ -148,7 +161,6 @@ class Application extends Singleton {
 					$current_user = null;
 					break;
 			}
-			return true;
 		} else {
 			require_once ABSPATH . '/wp-admin/includes/user.php';
 			/**
@@ -158,20 +170,44 @@ class Application extends Singleton {
 			 * This filter allows you to add conditional assignment.
 			 *
 			 * @filter nlmg_assign_to
-			 * @since 1.0.0
+			 *
 			 * @param int|string User ID to be assigned. Might be 0 or empty string.
 			 * @param int $user_id User ID to leave
+			 *
 			 * @return int|string
+			 * @since 1.0.0
 			 */
-			$assign_to = apply_filters( 'nlmg_assign_to', $this->option['assign_to'] ? $this->option['assign_to'] : null, $user_id );
+			$assign_to = apply_filters( 'nlmg_assign_to', $this->option[ 'assign_to' ] ? $this->option[ 'assign_to' ] : null, $user_id );
+			
 			return wp_delete_user( $user_id, $assign_to );
 		}
 	}
-
+	
+	/**
+	 * Delete user meta.
+	 *
+	 * @param int $user_id
+	 * @return int Deleted meta records.
+	 */
+	public function delete_user_meta( $user_id ) {
+		global $wpdb;
+		$keys = implode( ',', array_map( function( $key ) use ( $wpdb ) {
+			return $wpdb->prepare( '%s', $key );
+		}, $this->filtered_keys() ) );
+		$sql = <<<SQL
+			DELETE FROM {$wpdb->usermeta}
+			WHERE user_id = %d
+			  AND meta_key NOT IN ({$keys})
+SQL;
+		$sql = $wpdb->prepare( $sql, $user_id );
+		return (int) $wpdb->query( $sql );
+	}
+	
 	/**
 	 * Get redirect URL after remove account
 	 *
 	 * @param int $user_id
+	 *
 	 * @return string
 	 */
 	protected function default_redirect_link( $user_id = 0 ) {
@@ -179,14 +215,16 @@ class Application extends Singleton {
 		 * nlmg_redirect_link
 		 *
 		 * @filter nlmg_redirect_link
-		 * @since 1.0.0
+		 *
 		 * @param string $link Default is `wp_login_url()`
-		 * @param int    $user_id User ID removed.
+		 * @param int $user_id User ID removed.
+		 *
 		 * @return string
+		 * @since 1.0.0
 		 */
 		return apply_filters( 'nlmg_redirect_link', wp_login_url(), $user_id );
 	}
-
+	
 	/**
 	 * Get confirm label
 	 *
@@ -205,7 +243,36 @@ class Application extends Singleton {
 		 */
 		return apply_filters( 'nlmg_resign_confirm_label', __( 'Are you sure to delete account?', 'never-let-me-go' ), get_current_user_id() );
 	}
-
+	
+	/**
+	 * Get available meta keys.
+	 *
+	 * @return string[]
+	 */
+	public function available_meta_keys() {
+		global $wpdb;
+		$query = <<<SQL
+			SELECT meta_key, COUNT(umeta_id) as total
+			FROM {$wpdb->usermeta}
+			GROUP BY meta_key
+			LIMIT 100
+SQL;
+		$keys = [];
+		foreach ( $wpdb->get_results( $query ) as $row ) {
+			$keys[ $row->meta_key ] = $row->total;
+		}
+		return $keys;
+	}
+	
+	/**
+	 * Get filter keys.
+	 *
+	 * @return string[]
+	 */
+	public function filtered_keys() {
+		return apply_filters( 'nlmg_allowed_keys', $this->meta_to_keep );
+	}
+	
 	/**
 	 * Getter
 	 *
@@ -217,41 +284,38 @@ class Application extends Singleton {
 		switch ( $name ) {
 			case 'input':
 				return Input::getInstance();
-				break;
 			case 'nonce_key':
 				return "_{$this->name}_nonce";
-				break;
 			case 'option':
 				$option = get_option( $this->name . '_option', array() );
 				foreach (
 					array(
-						'enable'        => 0,
-						'resign_page'   => 0,
-						'assign_to'     => 0,
-						'keep_account'  => 0,
-						'destroy_level' => 1,
+						'enable'             => 0,
+						'resign_page'        => 0,
+						'assign_to'          => 0,
+						'keep_account'       => 0,
+						'destroy_level'      => 1,
+						'meta_to_keep'       => 'rich_editing',
+						'display_acceptance' => 0,
 					) as $key => $val
 				) {
 					if ( ! isset( $option[ $key ] ) ) {
 						$option[ $key ] = $val;
 					}
 				}
-
+				
 				return $option;
-				break;
+			case 'meta_to_keep':
+				return array_filter( explode( ',', $this->option['meta_to_keep'] ) );
 			case 'version':
 				return NLMG_VERSION;
-				break;
 			case 'dir':
 				return dirname( NLMG_BASE_FILE );
-				break;
 			case 'url':
 				return plugin_dir_url( NLMG_BASE_FILE );
-				break;
 			default:
 				// Do nothing
 				return null;
-				break;
 		}
 	}
 }
